feat: convert save-as-xlsx-action.js to ts
