+++ src/global.d.ts
@@ -1,0 +1,8 @@
+declare module '@neovici/nullxlsx' {
+	class NullXlsx {
+		constructor(filename: string);
+		addSheetFromData(data: unknown[][], sheetname: string): this;
+		generate(): ArrayBuffer;
+	}
+	export { NullXlsx };
+}
